requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.4
Pillow==5.3.0
six==1.11.0
scipy==1.1.0
torch==0.4.1
torchvision==0.2.1
Werkzeug==0.14.1